User input accepted
Added the ability for the user to add their own personal information.
Need to keep the envelope's formatting regardless of user information
entered.

Need to work on getting the envelope to line up exactly, has to do with
"int" rounding leaving the finishing | to be one short in some cases.